clients/python/requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.0
fastavro==1.8.4
idna==3.4
kafka-python==2.0.2
python-dotenv==0.20.0
requests==2.31.0
urllib3==2.0.7
